requirements.txt
djangorestframework
Django
drf-yasg